requirements.txt
# Core
pandas==1.5.3
numpy==1.24.0
scikit-learn==1.3.0
xgboost==1.7.6

# XAI
shap==0.42.0
lime==0.2.0

# Dashboard
streamlit==1.28.0

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0

# Cloud
boto3==1.28.0

# Utils
pyyaml==6.0
python-dotenv==1.0.0
